Cope with unexpected initial backend states
If the OS is booted via iSCSI, then the backend will already be in the
Connected state when the xenvif driver loads.  The backend is stable
in this state, so __FrontendPrepare() eventually times out waiting for
the backend state to change.

Fix by leading the backend through the Closing and Closed states if we
find it initially in the Connected (or Closing) states.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>